+++ node/mocks/contextFactory.ts
@@ -57,6 +57,9 @@ export function createContext<Ctx = Context>({
       apps: {
         getAppSettings: jest.fn().mockReturnValue(appSettings ?? {}),
       },
+      vbase: {
+        getJSON: jest.fn().mockResolvedValue({}),
+      }
     },
     vtex: {
       production: production ?? false,