+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Feature flag to call the productSearch and facets using the intsch api.
+
+### Changed 
+
+- Make productSearch and facets send a % of the traffic to our routes on intsch.
+
+### Fixed
+
+- Make the productIdentifier call the intsch when the feature flag is on.
+
 ## [1.86.0] - 2025-10-21
 
 ### Changed